CDS styles to sign in and sign up view [CV2-3998] (#1734)
* start to restyle the login screen for CDS styles

* continue to work on layout and styles for login and other auth pages

* move invited form to better directory

* update confirmation screens with CDS styles [CV2-3998]

* let email address in translated text be clickable

* update change password forms

* [full ci] Run all integration tests for this branch

* try to fix integration test [full ci]

* add some bem int prefixes to classnames

* fix logic on setting the correct error message on invited form

* remove testing code

* update error message setting logic in the same way

* fix typo found in PR review
