+++ localization/react-intl/src/app/components/UserPasswordChange.json
@@ -1,10 +1,5 @@
 [
   {
-    "id": "global.bestViewed",
-    "description": "Message shown to the user when they are using an unsupported browser",
-    "defaultMessage": "Best viewed with <a href=\"{chromedownload}\">Chrome for Desktop</a>."
-  },
-  {
     "id": "passwordChange.successTitle",
     "description": "Title message when the user password was successfully updated",
     "defaultMessage": "Password updated"
