+++ localization/react-intl/src/app/components/UserTosForm.json
