+++ localization/react-intl/src/app/components/login/InviteNewAccount.json
