+++ localization/react-intl/src/app/components/login/Login.json
@@ -1,0 +1,82 @@
+[
+  {
+    "id": "login.tosMissing",
+    "description": "Error message to tell the user they must agree to the app terms of service before continuing",
+    "defaultMessage": "You must agree to the Terms of Service and Privacy Policy"
+  },
+  {
+    "id": "login.title",
+    "description": "Header title for the sign in page",
+    "defaultMessage": "Sign in"
+  },
+  {
+    "id": "login.registerTitle",
+    "description": "Header title for the new user registration page",
+    "defaultMessage": "Register"
+  },
+  {
+    "id": "login.withGoogle",
+    "description": "Button label for the user to sign in to the app using their google credentials",
+    "defaultMessage": "Sign in with Google"
+  },
+  {
+    "id": "login.emailLogin",
+    "description": "Button label for the user to sign in to the app using their email address",
+    "defaultMessage": "Or, sign in with your email"
+  },
+  {
+    "id": "login.nameLabel",
+    "description": "Text field label for the user's name",
+    "defaultMessage": "Name"
+  },
+  {
+    "id": "login.emailLabel",
+    "description": "Text field label for the user's email address",
+    "defaultMessage": "Email"
+  },
+  {
+    "id": "login.passwordInputHint",
+    "description": "Text field label for the user's password",
+    "defaultMessage": "Password"
+  },
+  {
+    "id": "login.passwordLabel",
+    "description": "Text field description for password input telling the user it much be at least 8 characters long when signing up",
+    "defaultMessage": "Password (minimum 8 characters)"
+  },
+  {
+    "id": "login.otpAttemptLabel",
+    "description": "Text field label for the user's two-factor authentication token",
+    "defaultMessage": "Two-Factor Authentication Token"
+  },
+  {
+    "id": "login.passwordConfirmLabel",
+    "description": "Text field label for the to confirm their password",
+    "defaultMessage": "Password confirmation"
+  },
+  {
+    "id": "login.signIn",
+    "description": "Sign in button label",
+    "defaultMessage": "Sign in"
+  },
+  {
+    "id": "login.signUp",
+    "description": "Sign up button label",
+    "defaultMessage": "Sign up"
+  },
+  {
+    "id": "loginEmail.lostPassword",
+    "description": "Link for the user to initiate a password reset if they do not know it",
+    "defaultMessage": "Forgot your password?"
+  },
+  {
+    "id": "login.signUpLink",
+    "description": "Button label. Switches from 'logging in' to 'create new account' mode",
+    "defaultMessage": "Don't have an account? <strong>Sign up</strong>"
+  },
+  {
+    "id": "login.signInLink",
+    "description": "Button label. Switches from 'create new account' to 'logging in' mode",
+    "defaultMessage": "Already have an account? <strong>Sign in</strong>"
+  }
+]
