+++ localization/react-intl/src/app/components/login/UserPasswordReset.json
@@ -12,7 +12,17 @@
   {
     "id": "passwordReset.confirmedText",
     "description": "Confirmation text to tell the user what will happen if their password request request was valid",
-    "defaultMessage": "If this email address exists, you will receive an email from {adminEmail} with instructions to reset your password. Make sure it didn't wind up in your spam mailbox. If you aren't receiving our password reset emails, contact {supportEmail}."
+    "defaultMessage": "If this email address exists, you will receive an email from <a href=\"mailto:{adminEmail}\">{adminEmail}</a> with instructions to reset your password."
+  },
+  {
+    "id": "passwordReset.confirmedTextSpam",
+    "description": "Confirmation text to tell the user that password resets may end up in spam folders",
+    "defaultMessage": "Make sure it did not end up in your spam mailbox."
+  },
+  {
+    "id": "passwordReset.confirmedTextExtra",
+    "description": "Confirmation text to tell the user how to contact support if they do not receive a password reset email",
+    "defaultMessage": "If you are not receiving our password reset emails, contact <a href=\"mailto:{supportEmail}\">{supportEmail}</a>."
   },
   {
     "id": "passwordReset.signIn",
@@ -21,7 +31,7 @@
   },
   {
     "id": "passwordReset.text",
-    "description": "Helper text about why the user should enter their email adddress",
+    "description": "Helper text about why the user should enter their email address",
     "defaultMessage": "Add your address and an email will be sent with further instructions."
   },
   {
