+++ src/app/components/UserTosForm.js
@@ -1,11 +1,22 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedDate, FormattedMessage } from 'react-intl';
-import Box from '@material-ui/core/Box';
+import { FormattedDate, defineMessages, injectIntl, FormattedMessage } from 'react-intl';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import { units } from '../styles/js/shared';
 import { stringHelper } from '../customHelpers';
+
+const messages = defineMessages({
+  tos: {
+    id: 'userTos.tosLink',
+    defaultMessage: 'Terms of Service',
+    description: 'Link text to take the user to the terms of service',
+  },
+  pp: {
+    id: 'userTos.ppLink',
+    defaultMessage: 'Privacy Policy',
+    description: 'Link text to take the user to the privacy policy',
+  },
+});
 
 class UserTosForm extends Component {
   constructor(props) {
@@ -14,18 +25,14 @@ class UserTosForm extends Component {
   }
 
   render() {
-    const linkStyle = {
-      textDecoration: 'underline',
-    };
-
     const tosLink = (
       <a
         target="_blank"
         rel="noopener noreferrer"
-        style={linkStyle}
         href={stringHelper('TOS_URL')}
+        title={this.props.intl.formatMessage(messages.tos)}
       >
-        <FormattedMessage id="userTos.tosLink" defaultMessage="Terms of Service" description="Link text to take the user to the terms of service" />
+        {this.props.intl.formatMessage(messages.tos)}
       </a>
     );
 
@@ -33,55 +40,51 @@ class UserTosForm extends Component {
       <a
         target="_blank"
         rel="noopener noreferrer"
-        style={linkStyle}
         href={stringHelper('PP_URL')}
+        title={this.props.intl.formatMessage(messages.pp)}
       >
-        <FormattedMessage id="userTos.ppLink" defaultMessage="Privacy Policy" description="Link text to take the user to the privacy policy" />
+        {this.props.intl.formatMessage(messages.pp)}
       </a>
     );
 
     const { termsLastUpdatedAt } = this.props;
 
     return (
-      <div>
+      <>
         { !this.props.user.last_accepted_terms_at ?
-          <div>
+          <>
             { this.props.showTitle ?
-              <h2 className="typography-h6">
+              <h6>
                 <FormattedMessage
                   id="userTos.title"
                   defaultMessage="Terms of Service and Privacy Policy"
                   description="Page title for the terms of service"
                 />
-              </h2> : null
+              </h6> : null
             }
             { termsLastUpdatedAt ?
-              <p className="typography-caption" style={{ margin: `${units(1)} 0` }}>
-                <FormattedMessage
-                  id="userTos.termsLastUpdatedAt"
-                  defaultMessage="Last updated {lastUpdated}"
-                  description="Date of the last terms of service update"
-                  values={{
-                    lastUpdated: <FormattedDate value={termsLastUpdatedAt * 1000} day="numeric" month="long" year="numeric" />,
-                  }}
-                />
-              </p> : null
+              <FormattedMessage
+                tagName="p"
+                id="userTos.termsLastUpdatedAt"
+                defaultMessage="Last updated {lastUpdated}"
+                description="Date of the last terms of service update"
+                values={{
+                  lastUpdated: <FormattedDate value={termsLastUpdatedAt * 1000} day="numeric" month="long" year="numeric" />,
+                }}
+              /> : null
             }
-            <Box my={4}>
-              <div className="typography-body1">
-                <FormattedMessage
-                  id="userTos.disclaimer"
-                  defaultMessage="Please review our {tosLink} and our {ppLink} and consent to the following:"
-                  description="Message for the user to review the terms of service and privacy policy"
-                  values={{
-                    tosLink,
-                    ppLink,
-                  }}
-                />
-              </div>
-            </Box>
-          </div> :
-          <div>
+            <FormattedMessage
+              tagName="p"
+              id="userTos.disclaimer"
+              defaultMessage="Please review our {tosLink} and our {ppLink} and consent to the following:"
+              description="Message for the user to review the terms of service and privacy policy"
+              values={{
+                tosLink,
+                ppLink,
+              }}
+            />
+          </> :
+          <>
             <h2 className="typography-subtitle2">
               <FormattedMessage
                 id="userTos.titleUpdated"
@@ -89,38 +92,36 @@ class UserTosForm extends Component {
                 description="Page title for the updated terms of service and privacy policy messages"
               />
             </h2>
-            <p style={{ margin: `${units(4)} 0` }}>
-              <FormattedMessage
-                id="userTos.disclaimerUpdate"
-                defaultMessage="We've updated our {tosLink} and our {ppLink}. Please review and consent to the following:"
-                description="Message to the user that the terms of service and privacy policy have been updated"
-                values={{
-                  tosLink,
-                  ppLink,
-                }}
-              />
-            </p>
-          </div>
+            <FormattedMessage
+              tagName="p"
+              id="userTos.disclaimerUpdate"
+              defaultMessage="We've updated our {tosLink} and our {ppLink}. Please review and consent to the following:"
+              description="Message to the user that the terms of service and privacy policy have been updated"
+              values={{
+                tosLink,
+                ppLink,
+              }}
+            />
+          </>
         }
-        <div style={{ margin: `${units(4)} 0` }}>
-          <FormControlLabel
-            control={
-              <Checkbox
-                id="tos__tos-agree"
-                onChange={this.props.handleCheckTos}
-                checked={this.props.checkedTos}
-              />
-            }
-            label={
-              <FormattedMessage
-                id="userTos.agreeTos"
-                defaultMessage="I agree to the Terms of Service."
-                description="Checkbox label for the user to agree to the terms of service"
-              />
-            }
-          />
-        </div>
-      </div>
+        <FormControlLabel
+          control={
+            <Checkbox
+              id="tos__tos-agree"
+              onChange={this.props.handleCheckTos}
+              checked={this.props.checkedTos}
+            />
+          }
+          label={
+            <FormattedMessage
+              tagName="p"
+              id="userTos.agreeTos"
+              defaultMessage="I agree to the Terms of Service."
+              description="Checkbox label for the user to agree to the terms of service"
+            />
+          }
+        />
+      </>
     );
   }
 }
@@ -139,4 +140,4 @@ UserTosForm.defaultProps = {
   showTitle: false,
 };
 
-export default UserTosForm;
+export default injectIntl(UserTosForm);
